Move javascript for video editor into its own file as well
