+++ dashboard/app/views/levels/editors/fields/_video.html.haml
@@ -1,8 +1,8 @@
 - content_for :body_scripts do
   - video_infos = {}
   - Video.all.each { |video| video_infos[video.key] = video.summarize(false) }
-  - script_data = { editorall: { video_infos: video_infos }.to_json }
-  %script{src: webpack_asset_path('js/levels/editors/_all.js'), data: script_data}
+  - script_data = { editorvideo: { video_infos: video_infos }.to_json }
+  %script{src: webpack_asset_path('js/levels/editors/_video.js'), data: script_data}
 
 .field
   %h3 Video
